+++ daiv/automation/agents/pipeline_fixer/conf.py
@@ -14,17 +14,17 @@ class PipelineFixerSettings(BaseSettings):
 
     NAME: str = Field(default="PipelineFixer", description="Name of the pipeline fixer agent.")
     MAX_ITERATIONS: int = Field(default=20, description="Maximum number of retry iterations for pipeline fixer.")
-    LOG_EVALUATOR_MODEL_NAME: ModelName | str = Field(
-        default=ModelName.CLAUDE_3_5_HAIKU, description="Model name to be used for log evaluator."
+    SAME_ERROR_MODEL_NAME: ModelName | str = Field(
+        default=ModelName.GPT_4_1_MINI, description="Model name to be used for same error evaluator."
     )
     TROUBLESHOOTING_MODEL_NAME: ModelName | str = Field(
-        default=ModelName.O3_MINI, description="Model name to be used for pipeline fixer."
+        default=ModelName.O4_MINI, description="Model name to be used for pipeline fixer."
     )
     TROUBLESHOOTING_THINKING_LEVEL: ThinkingLevel = Field(
         default=ThinkingLevel.HIGH, description="Thinking level to be used for pipeline fixer."
     )
-    LINT_EVALUATOR_MODEL_NAME: ModelName | str = Field(
-        default=ModelName.CLAUDE_3_5_HAIKU, description="Model name to be used for lint evaluator."
+    COMMAND_OUTPUT_MODEL_NAME: ModelName | str = Field(
+        default=ModelName.GPT_4_1_MINI, description="Model name to be used for command output evaluator."
     )