+++ daiv/automation/agents/review_addressor/agent.py
@@ -7,7 +7,7 @@ from django.utils import timezone
 
 from langchain_core.output_parsers.openai_tools import PydanticToolsParser
 from langchain_core.prompts import ChatPromptTemplate, MessagesPlaceholder, SystemMessagePromptTemplate
-from langchain_core.runnables import RunnableConfig, RunnablePassthrough
+from langchain_core.runnables import Runnable, RunnableConfig
 from langgraph.graph import END, StateGraph
 from langgraph.graph.state import CompiledStateGraph
 from langgraph.prebuilt import create_react_agent
@@ -22,17 +22,34 @@ from codebase.indexes import CodebaseIndex
 from core.config import RepositoryConfig
 
 from .conf import settings
-from .prompts import (
-    respond_reviewer_system,
-    review_assessment_human,
-    review_assessment_system,
-    review_plan_system_template,
-)
-from .schemas import ReviewAssessment
+from .prompts import respond_reviewer_system, review_comment_human, review_comment_system, review_plan_system_template
+from .schemas import ReviewCommentEvaluation, ReviewCommentInput
 from .state import OverallState, ReplyAgentState
 from .tools import reply_reviewer_tool
 
 logger = logging.getLogger("daiv.agents")
+
+
+class ReviewCommentEvaluator(BaseAgent[Runnable[ReviewCommentInput, ReviewCommentEvaluation]]):
+    """
+    Agent to evaluate if a review comment is a request for changes to the codebase.
+    """
+
+    def compile(self) -> Runnable:
+        return (
+            ChatPromptTemplate.from_messages([
+                review_comment_system,
+                MessagesPlaceholder("messages"),
+                review_comment_human,
+            ])
+            # We could use `with_structured_output` but it define tool_choice as "any", forcing the llm to respond with
+            # a tool call without reasoning, which is crucial here to make the right decision.
+            # Defining tool_choice as "auto" would let the llm to reason before calling the tool.
+            | self.get_model(model=settings.REVIEW_COMMENT_MODEL_NAME).bind_tools(
+                [ReviewCommentEvaluation], tool_choice="auto"
+            )
+            | PydanticToolsParser(tools=[ReviewCommentEvaluation], first_tool_only=True)
+        ).with_config({"run_name": "ReviewCommentEvaluator"})
 
 
 class ReviewAddressorAgent(BaseAgent[CompiledStateGraph]):
@@ -74,23 +91,10 @@ class ReviewAddressorAgent(BaseAgent[CompiledStateGraph]):
         Returns:
             Command[Literal["plan_and_execute", "reply_reviewer"]]: The next step in the workflow.
         """
-        evaluator = (
-            RunnablePassthrough.assign(
-                messages=lambda inputs: inputs["messages"][:-1], comment=lambda inputs: inputs["messages"][-1].content
-            )
-            | ChatPromptTemplate.from_messages([
-                review_assessment_system,
-                MessagesPlaceholder("messages"),
-                review_assessment_human,
-            ])
-            # We could use `with_structured_output` but it define tool_choice as "any", forcing the llm to respond with
-            # a tool call without reasoning, which is crucial here to make the right decision.
-            # Defining tool_choice as "auto" would let the llm to reason before calling the tool.
-            | self.get_model(model=settings.ASSESSMENT_MODEL_NAME).bind_tools([ReviewAssessment], tool_choice="auto")
-            | PydanticToolsParser(tools=[ReviewAssessment], first_tool_only=True)
-        )
-
-        response = cast("ReviewAssessment", evaluator.invoke({"messages": state["notes"]}))
+        response = ReviewCommentEvaluator().agent.invoke({
+            "messages": state["notes"][:-1],
+            "comment": cast("str", state["notes"][-1].content),
+        })
 
         if response.request_for_changes:
             return Command(goto="plan_and_execute")
@@ -162,7 +166,11 @@ class ReviewAddressorAgent(BaseAgent[CompiledStateGraph]):
             version="v2",
         )
 
-        result = react_agent.invoke({"messages": state["notes"], "diff": state["diff"]})
+        result = react_agent.invoke({
+            "messages": state["notes"],
+            "diff": state["diff"],
+            "current_date_time": timezone.now().strftime("%d %B, %Y %H:%M"),
+        })
 
         # The reply is updated in the state by the reply_reviewer tool.
         # There's cases where the tool is not called, so we use the last message content as the reply.